% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{host_example}
\alias{host_example}
\title{dRNA-seq read count matrix from the host with 6 time points
(0 h, 2 h, 4 h, 8 h, 16 h, and 24 h)}
\format{
a data frame with 40375 rows and 18 variables
\describe{
\item{HeLa.S3.WT.00_h.R1}{round 1 hour 0 read counts}
\item{HeLa.S3.WT.00_h.R2}{round 2 hour 0 read counts}
\item{HeLa.S3.WT.00_h.R3}{round 3 hour 0 read counts}
\item{HeLa.S3.WT.02_h.R1}{round 1 hour 2 read counts}
\item{HeLa.S3.WT.02_h.R2}{round 2 hour 2 read counts}
\item{HeLa.S3.WT.02_h.R3}{round 3 hour 2 read counts}
\item{HeLa.S3.WT.04_h.R1}{round 1 hour 4 read counts}
\item{HeLa.S3.WT.04_h.R2}{round 2 hour 4 read counts}
\item{HeLa.S3.WT.04_h.R3}{round 3 hour 4 read counts}
\item{HeLa.S3.WT.08_h.R1}{round 1 hour 8 read counts}
\item{HeLa.S3.WT.08_h.R2}{round 2 hour 8 read counts}
\item{HeLa.S3.WT.08_h.R3}{round 3 hour 8 read counts}
\item{HeLa.S3.WT.16_h.R1}{round 1 hour 16 read counts}
\item{HeLa.S3.WT.16_h.R2}{round 2 hour 16 read counts}
\item{HeLa.S3.WT.16_h.R3}{round 3 hour 16 read counts}
\item{HeLa.S3.WT.24_h.R1}{round 1 hour 24 read counts}
\item{HeLa.S3.WT.24_h.R2}{round 2 hour 24 read counts}
\item{HeLa.S3.WT.24_h.R3}{round 3 hour 24 read counts}
}
}
\source{
NCBI’s Gene Expression Omnibus (accession number GSE60144)
}
\usage{
host_example
}
\description{
dRNA-seq read count matrix from the host with 6 time points
(0 h, 2 h, 4 h, 8 h, 16 h, and 24 h)
}
\examples{
\dontrun{
norm_TMM
identifyDE
volcanoPlot

}

}
\keyword{datasets}
