% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialExpression.R
\name{identifyDE}
\alias{identifyDE}
\title{Separate data by categorizing them by hour groups (ie. WT.02_h, WT.04_h, etc.) to analyze
differential expression at different time periods}
\usage{
identifyDE(data, pheno)
}
\arguments{
\item{data}{dRNA-seq tab-delimited input csv file with read counts}

\item{pheno}{csv file containing sample names and time/condition of each sample}
}
\value{
list of differentially expressed genes compared with the baseline (0 hours)
}
\description{
The limma package is used to find differentially expressed (DE) genes, in which expression data
is described by linear modelling
}
\examples{
\dontrun{
norm <- norm_TMM(host_example, phenotype_example)
DE <- identifyDE(norm, phenotype_example)
}

}
\references{
Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK (2015).
“limma powers differential expression analyses for RNA-sequencing and
microarray studies.” Nucleic Acids Research, 43(7), e47. doi: 10.1093/nar/gkv007

Dinarvand, M., Kock, F., Al Mouiee, D., Vuong, K., Vijayan, A., Tanzim, A.
F., Azad, A. K. M., Penesyan, A., Castaño-Rodríguez, N., & Vafaee, F. (2022).
DSEQSB: A Systems Biology Approach to decipher dynamics of host-pathogen
interactions using temporal dual RNA-Seq Data. https://doi.org/10.1101/2022.02.28.482417

Macho Rendón, J., Lang, B., Ramos Llorens, M., Tartaglia, G.G., and
Torrent Burgas, M. (2021). DualSeqDB: a database to assess the relevance of
bacterial genes during host infection. Nucleic Acids Res. 49, D687–D693.
}
