% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_TMM}
\alias{norm_TMM}
\title{Normalize data by using trimmed mean of M-values to convert size of data to an effective library size}
\usage{
norm_TMM(data, pheno)
}
\arguments{
\item{data}{dRNA-seq tab-delimited input data frame with read counts}

\item{pheno}{data frame containing sample names and time/condition of each sample}
}
\value{
data.norm DGEList
}
\description{
Normalize data by using trimmed mean of M-values to convert size of data to an effective library size
}
\examples{
\dontrun{
 norm_TMM(host_example, phenotype_example)
}
}
\references{
Robinson MD, McCarthy DJ and Smyth GK (2010). edgeR: a Bioconductor package
for differential expression analysis of digital gene expression data.
Bioinformatics 26, 139-140
}
