% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotype_example}
\alias{phenotype_example}
\title{phenotype data frame that contains names of the columns headers corresponding
expression (read count) data frame, and group name which is time/condition}
\format{
a data frame with 18 rows and 1 variables
\describe{
\item{groups}{time or condition}
}
}
\source{
NCBI’s Gene Expression Omnibus (accession number GSE60144)
}
\usage{
phenotype_example
}
\description{
phenotype data frame that contains names of the columns headers corresponding
expression (read count) data frame, and group name which is time/condition
}
\examples{
\dontrun{
norm_TMM
identifyDE
volcanoPlot
}

}
\keyword{datasets}
