% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read_csv}
\alias{read_csv}
\title{Read csv files to access data}
\usage{
read_csv(path)
}
\arguments{
\item{path}{file path}
}
\value{
data frame (data.frame) containing a representation of the data in the file
}
\description{
Read csv files to access data
}
