% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRnaDualSeq.R
\name{runRnaDualSeq}
\alias{runRnaDualSeq}
\title{Launch Shiny App for rnaDualSeq}
\usage{
runRnaDualSeq()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for rnaDualSeq
}
\examples{
\dontrun{

rnaDualSeq::runRnaDualSeq()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
