% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Creates 5 volcano plots denoting each time period in data set visualing
significant upregulation and downregulation genes}
\usage{
volcanoPlot(
  name,
  de_norm,
  timeperiods = c("2h", "4h", "8h", "16h", "24h"),
  hour = "2h"
)
}
\arguments{
\item{name}{what gene is being graphed}

\item{de_norm}{A nested list containing differentially expressed results}

\item{timeperiods}{a vector containing labels for each time period}

\item{hour}{time period for graph}
}
\value{
graph at time period of volcano plot saves pngs images
into a results folder
}
\description{
Creates 5 volcano plots denoting each time period in data set visualing
significant upregulation and downregulation genes
}
\examples{
\dontrun{
norm <- norm_TMM(host_example, phenotype_example)
DE <- identifyDE(norm, phenotype_example)
volcanoPlot("Host Example", DE)
}

}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4, https://ggplot2.tidyverse.org.
}
