% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialExpression.R
\name{identifyDE}
\alias{identifyDE}
\title{Separate data by categorizing them by hour groups (ie. WT.02_h, WT.04_h, etc.) to analyze
differential expression at different time periods}
\usage{
identifyDE(data, pheno)
}
\arguments{
\item{data}{dRNA-seq tab-delimited input csv file with read counts}

\item{pheno}{csv file containing sample names and time/condition of each sample}
}
\value{
list of differentially expressed genes compared with the baseline (0 hours)
}
\description{
The limma package is used to find differentially expressed (DE) genes, in which expression data
is described by linear modelling
}
\examples{
\dontrun{
norm <- norm_TMM(host_example, phenotype_example)
DE <- identifyDE(norm, phenotype_example)
}

}
