% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_TMM}
\alias{norm_TMM}
\title{Normalize data by using trimmed mean of M-values to convert size of data to an effective library size}
\usage{
norm_TMM(data, pheno)
}
\arguments{
\item{data}{dRNA-seq tab-delimited input csv file with read counts}

\item{pheno}{csv file containing sample names and time/condition of each sample}
}
\value{
data.norm DGEList
}
\description{
Normalize data by using trimmed mean of M-values to convert size of data to an effective library size
}
