% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Creates 5 volcano plots denoting each time period in data set visualing
significant upregulation and downregulation genes}
\usage{
volcanoPlot(
  name,
  differentialExpression,
  timeperiods = c("2h", "4h", "8h", "16h", "24h")
)
}
\arguments{
\item{name}{what gene is being graphed}

\item{differentialExpression}{A nested list containing differentially expressed results}

\item{timeperiods}{a vector containing labels for each time period}
}
\value{
images of volcano plots per time period
}
\description{
Creates 5 volcano plots denoting each time period in data set visualing
significant upregulation and downregulation genes
}
